REFACTOR(RestApi proj): extracted base controller for non-CRUD purposes
