+++ OnlineMinion.RestApi/BaseControllers/BaseCRUDApiController.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel.DataAnnotations;
 using System.Globalization;
-using ErrorOr;
 using MediatR;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
@@ -11,11 +10,9 @@ using OnlineMinion.RestApi.Configuration;
 using Swashbuckle.AspNetCore.Annotations;
 using Swashbuckle.AspNetCore.Filters;
 
-namespace OnlineMinion.RestApi;
+namespace OnlineMinion.RestApi.BaseControllers;
 
-[Route("api/v{version:apiVersion}/[controller]")]
-[ApiController]
-public abstract class ApiControllerBase(ISender sender) : ControllerBase
+public abstract class BaseCRUDApiController(ISender sender) : BaseApiController
 {
     protected readonly ISender Sender = sender;
 
@@ -81,51 +78,6 @@ public abstract class ApiControllerBase(ISender sender) : ControllerBase
         headers[CustomHeaderNames.PagingSize] = values.Size.ToString(NumberFormatInfo.InvariantInfo);
         headers[CustomHeaderNames.PagingPages] = values.Pages.ToString(NumberFormatInfo.InvariantInfo);
     }
-
-    /// <summary>
-    ///     Sets ModelState and "converts" <see cref="Error" /> to <see cref="IActionResult" />.
-    /// </summary>
-    /// <param name="error">Error to convert to action result.</param>
-    /// <param name="instanceUrl">Optional URL similar to <code>GET /api/[controller]/{id}</code>.</param>
-    protected ActionResult CreateApiProblemResult(Error error, string? instanceUrl = null)
-    {
-        if (error.Dictionary is not null)
-        {
-            SaveToModelState(error.Dictionary);
-        }
-
-        return error.Type switch
-        {
-            ErrorType.Conflict => ValidationProblem(error.Description, statusCode: StatusCodes.Status409Conflict),
-            ErrorType.Validation => ValidationProblem(error.Description, instanceUrl),
-            _ => Problem(error.Description, title: error.Code, instance: instanceUrl),
-        };
-    }
-
-    private void SaveToModelState(IDictionary<string, object> errorMetadata)
-    {
-        foreach (var (key, data) in errorMetadata)
-        {
-            var values = data as IEnumerable<string>
-                         ?? new[] { data.ToString() ?? string.Empty, };
-
-            foreach (var v in values)
-            {
-                ModelState.AddModelError(
-                    key,
-                    string.IsNullOrWhiteSpace(v) ? "<Missing error message>" : v
-                );
-            }
-        }
-    }
-
-    /// <inheritdoc cref="UrlHelperExtensions.Action(Microsoft.AspNetCore.Mvc.IUrlHelper)" />
-    /// <param name="getByIdActionName">
-    ///     Name of action, one that produces similar to <code>GET /api/[controller]/{id}</code>.
-    /// </param>
-    /// <param name="id">Id value that came from route parameter.</param>
-    protected virtual string? GetInstanceUrl(string getByIdActionName, int id) =>
-        Url.Action(getByIdActionName, new { id, });
 
     /// <summary>
     ///     Checks id equality between route parameter and request body.
