+++ OnlineMinion.RestApi/TransactionCreditsController.cs
@@ -9,6 +9,7 @@ using OnlineMinion.Contracts.Shared.Requests;
 using OnlineMinion.Contracts.Shared.Responses;
 using OnlineMinion.Contracts.Transactions.Credit.Requests;
 using OnlineMinion.Contracts.Transactions.Credit.Responses;
+using OnlineMinion.RestApi.BaseControllers;
 using OnlineMinion.RestApi.Configuration;
 using Swashbuckle.AspNetCore.Annotations;
 using Swashbuckle.AspNetCore.Filters;
@@ -18,7 +19,7 @@ namespace OnlineMinion.RestApi;
 [Route("api/v{version:apiVersion}/Transactions/Credits")]
 [ApiVersion("1")]
 public class TransactionCreditsController(ISender sender, ILogger<TransactionCreditsController> logger)
-    : ApiControllerBase(sender)
+    : BaseCRUDApiController(sender)
 {
     [HttpGet("{id}")]
     [ProducesResponseType<TransactionCreditResp>(StatusCodes.Status200OK)]
